VF vfpminifire_update: hoist delta subtractions + signed random compares + pointer null-checks (91.2->96.8)

+++ src/main/dll/dll_0000_baby_snowworm.c
@@ -2347,7 +2347,7 @@ void fn_80128470(int p1)
     gameTextSetColor(0xff, 0xff, 0xff,
                      (int)((double)((s16)p1 * (0x200 - lbl_803DD75C)) * lbl_803E2088));
     lbl_803DBA8A = (s16)(0x100 - lbl_803DD75C);
-    if (pauseMenuState < 0xb && pauseMenuState >= 8)
+    if ((s8)pauseMenuState < 0xb && (s8)pauseMenuState >= 8)
     {
         gameTextFn_80016810(0x3e8, 0xc8, 0x154);
     }
